requirements_docker.txt
colorlog==4.8.0
deepface==0.0.95
google-api-core==2.28.1
google-auth==2.41.1
google-api-python-client==2.187.0
gunicorn==20.1.0
matplotlib==3.10.7
mediapipe==0.10.21
numpy==1.24.3
pandas==2.3.3
protobuf==4.25.3
python-dotenv==1.2.1
PyYAML==6.0.3
tensorflow==2.13.1
tomli==2.3.0
faiss-cpu
fastapi
uvicorn==0.38.0
python-multipart==0.0.20
fastparquet==2024.11.0
python-json-logger
onnxruntime
psycopg2-binary==2.9.11
pyzmq==27.1.0

